pyproject.toml
[project]
name = "art-e"
version = "0.1.0"
description = "Deep learning research for email search and processing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.4.1",
    "diskcache>=5.6.3",
    "huggingface-hub>=0.29.3",
    "langchain-core>=0.3.51",
    "litellm>=1.65.0.post1",
    "matplotlib>=3.10.1",
    "openpipe-art[backend]",
    "pandas>=1.3.0",
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.62.0",
    "transformers>=4.50.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["art_e*"]

[tool.uv.sources]
art-e = { workspace = true }

[dependency-groups]
dev = [
    "art-e",
    "pytest>=8.3.5",
    "ruff>=0.8",
]
